third_party/blink/web_tests/platform/generic/touchadjustment/nested-shadow-node-expected.txt

Test the case where a clickable target contains a shadow-DOM element. The adjusted point should snap to the location of the shadow-DOM element if close enough to the original touch position.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS adjusted node was DIV#targetDiv.
PASS adjustedPoint.x is within 10 of 88
PASS adjustedPoint.y is within 10 of 28
PASS adjusted node was DIV#targetDiv.
PASS adjustedPoint.x is within 10 of 88
PASS adjustedPoint.y is within 10 of 28
PASS adjusted node was DIV#targetDiv.
PASS adjustedPoint.x is within 10 of 88
PASS adjustedPoint.y is within 10 of 28
PASS adjusted node was DIV#targetDiv.
PASS adjustedPoint.x is within 10 of 88
PASS adjustedPoint.y is within 10 of 28
PASS adjusted node was DIV#targetDiv.
PASS adjustedPoint.x is within 1 of 58
PASS adjustedPoint.y is within 1 of 58
PASS successfullyParsed is true

TEST COMPLETE
